reorganize components and folders
