+++ src/components/App.jsx
@@ -9,11 +9,15 @@ import { NavBar } from './AppStyles.jsx';
 
 const App = () => {
   const [currentItem, setCurrentItem] = useState({});
+  const [productId, setProductId] = useState('');
 
   const getFirstItem = () => {
     fetch(`${process.env.API_URI}/products`, { method: 'GET', headers: { Authorization: process.env.API_KEY } })
       .then((response) => {
-        response.json().then((results) => setCurrentItem(results[0]));
+        response.json().then((results) => {
+          setCurrentItem(results[0]);
+          setProductId(results[0].id);
+        });
       })
       .catch((err) => {
         console.log(`Error found in getFirstItem: ${err}`);
@@ -29,7 +33,7 @@ const App = () => {
       <NavBar>Kids Next Door</NavBar>
       <Overview currentItem={currentItem}/>
       <RelatedList />
-      <QuestionsAnswers />
+      <QuestionsAnswers id={productId} />
       <ReviewList />
     </>
   )
