+++ src/components/questionsAnswers/Styles.jsx
@@ -1,0 +1,10 @@
+/* eslint-disable */
+import styled from 'styled-components';
+
+export const Main = styled.div`
+  margin: 1em;
+  padding: 1em;
+  border: 2px solid #000;
+  display: flex;
+  flex-flow: column;
+`;
