+++ src/components/questionsAnswers/answers/Styles.jsx
@@ -1,0 +1,6 @@
+/* eslint-disable */
+import styled from 'styled-components';
+
+export const AnswerList = styled.ul`
+  list-style: none;
+`;
