+++ src/components/questionsAnswers/questions/Styles.jsx
@@ -1,0 +1,24 @@
+/* eslint-disable */
+import styled from 'styled-components';
+
+export const List = styled.div(props => ({
+  border: '2px solid #000',
+  height: props.height,
+  overflow: 'auto'
+}))
+
+export const Question = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0 1em;
+`
+
+export const FormInner = styled.div`
+  display: flex;
+  flex-flow: column;
+  justify-content: center;
+  align-items: center;
+  margin: auto;
+  padding: 0.5em;
+`
