+++ src/components/questionsAnswers/search/Styles.jsx
@@ -1,0 +1,7 @@
+/* eslint-disable */
+import styled from 'styled-components';
+
+export const SearchInput = styled.input`
+  padding: 5px;
+  margin: 0.5em 0;
+`;
